+++ device/google/contexthub/firmware/variant/nucleo/inc/variant/variant.h
@@ -42,13 +42,13 @@ extern "C" {
 
 #define BMI160_TO_ANDROID_COORDINATE(x, y, z)   \
     do {                                        \
-        int32_t xi = x, yi = y, zi = z;         \
+        float xi = x, yi = y, zi = z;           \
         x = xi; y = yi; z = zi;                 \
     } while (0)
 
 #define BMM150_TO_ANDROID_COORDINATE(x, y, z)   \
     do {                                        \
-        int32_t xi = x, yi = y, zi = z;         \
+        float xi = x, yi = y, zi = z;           \
         x = xi; y = -yi; z = -zi;               \
     } while (0)
